+++ users/views/lists.py
@@ -1,0 +1,160 @@
+from rest_framework import viewsets, status
+from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
+from drf_spectacular.utils import extend_schema, OpenApiParameter, OpenApiTypes
+from django.shortcuts import get_object_or_404
+from users.models import Profile, Skill, Language, Badge
+from users.serializers import UsersBySkillSerializer, UsersByTagSerializer, ProfileSerializer
+from users.submodels import Territory, WikimediaProject
+from orgs.models import Organization
+from events.models import Events
+from projects.models import Project
+
+
+class UsersBySkillViewSet(viewsets.ReadOnlyModelViewSet):
+    queryset = Profile.objects.all()
+    serializer_class = UsersBySkillSerializer
+
+    @extend_schema(
+        summary='List users by skill.',
+        description='Deprecated. This endpoint lists users by skill. Please use the /tags/ endpoint instead.',
+        deprecated=True
+    )
+    def retrieve(self, request, *args, **kwargs):
+        skill_id = self.kwargs['pk']
+        skill = get_object_or_404(Skill, pk=skill_id)
+
+        known_users = Profile.objects.filter(skills_known=skill)
+        available_users = Profile.objects.filter(skills_available=skill)
+        wanted_users = Profile.objects.filter(skills_wanted=skill)
+        data = {
+            'known': [{'id': user.id, 'display_name': user.display_name, 'username': user.user.username, 'profile_image': user.profile_image} for user in known_users],
+            'available': [{'id': user.id, 'display_name': user.display_name, 'username': user.user.username, 'profile_image': user.profile_image} for user in available_users],
+            'wanted': [{'id': user.id, 'display_name': user.display_name, 'username': user.user.username, 'profile_image': user.profile_image} for user in wanted_users],
+        }
+        return Response(data)
+
+    @extend_schema(exclude=True)
+    def list(self, request, *args, **kwargs):
+        response = {'message': 'Please provide a skill id.'}
+        return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+
+class QuickListViewSet(viewsets.ReadOnlyModelViewSet):
+    lookup_url_kwarg = 'list_type'
+
+    def get_queryset(self):
+        list_type = self.kwargs.get('list_type')
+        if list_type == 'language':
+            return Language.objects.all()
+        elif list_type == 'wikimedia_project':
+            return WikimediaProject.objects.all()
+        elif list_type == 'affiliation':
+            return Organization.objects.all()
+        elif list_type == 'territory':
+            return Territory.objects.all()
+        elif list_type == 'event':
+            return Events.objects.all()
+        elif list_type == 'project':
+            return Project.objects.all()
+        elif list_type == 'skills':
+            return Skill.objects.all()
+        elif list_type == 'badges':
+            return Badge.objects.all()
+        elif list_type == 'users':
+            return Profile.objects.all()
+        else:
+            # Dummy empty queryset to avoid errors in the schema generation
+            return Profile.objects.none()
+    
+    def get_serializer_class(self): # pragma: no cover
+        # Dummy method to avoid errors in the schema generation
+        return ProfileSerializer
+
+    @extend_schema(
+        summary='List all items in a simplified way.',
+        description='This endpoint lists all items of a given type in a simplified way.',
+        parameters=[
+            OpenApiParameter(
+                "list_type",
+                OpenApiTypes.STR,
+                OpenApiParameter.PATH,
+                required=True,
+                description='The type of list to retrieve.',
+                enum=['language', 'wikimedia_project', 'affiliation', 'territory', 'skills', 'event', 'project', 'badges', 'users'],
+            ),
+        ],
+        responses={(200, 'application/json'): {
+            'description': 'A mapping of item IDs to item names.',
+            'type': 'object',
+            'additionalProperties': {
+                'type': 'string',
+            },
+            'example': {
+                '1': 'Label 1',
+                '2': 'Label 2',
+                '3': 'Label 3',
+            },
+        }},
+    )
+    def retrieve(self, request, *args, **kwargs):
+        queryset = self.get_queryset()
+        data = {item.id: str(item) for item in queryset}
+        return Response(data)
+
+    @extend_schema(
+        exclude=True
+    )
+    def list(self, request, *args, **kwargs):
+        return Response({'message': 'Please provide a valid list type.'}, status=status.HTTP_400_BAD_REQUEST)
+
+
+class UsersByTagViewSet(viewsets.ReadOnlyModelViewSet):
+    queryset = Profile.objects.all()
+    serializer_class = UsersByTagSerializer
+
+    @extend_schema(
+        summary='Lists users by tag.',
+        description='This endpoint retrieves users by tag.',
+        parameters=[
+            OpenApiParameter(
+                "tag_type",
+                OpenApiTypes.STR,
+                OpenApiParameter.PATH,
+                required=True,
+                description='The type of tag to search for.',
+                enum=['skill_known', 'skill_available', 'skill_wanted', 'language', 'territory', 'wikimedia_project', 'affiliation'],
+            ),
+            OpenApiParameter(
+                "tag_id",
+                OpenApiTypes.INT,
+                OpenApiParameter.PATH,
+                required=True,
+                description='The ID of the tag to search for.',
+            ),
+        ],
+    )
+    def list(self, request, *args, **kwargs):
+        tag_type = kwargs.get('tag_type')
+        tag_id = kwargs.get('tag_id')
+        if not tag_type or not tag_id:
+            return Response({'message': 'Please provide a valid tag type and tag ID.'}, status=status.HTTP_400_BAD_REQUEST)
+
+        if tag_type == 'skill_known':
+            queryset = Profile.objects.filter(skills_known__id=tag_id)
+        elif tag_type == 'skill_available':
+            queryset = Profile.objects.filter(skills_available__id=tag_id)
+        elif tag_type == 'skill_wanted':
+            queryset = Profile.objects.filter(skills_wanted__id=tag_id)
+        elif tag_type == 'language':
+            queryset = Profile.objects.filter(languageproficiency__language__id=tag_id)
+        elif tag_type == 'territory':
+            queryset = Profile.objects.filter(territory__id=tag_id)
+        elif tag_type == 'wikimedia_project':
+            queryset = Profile.objects.filter(wikimedia_project__id=tag_id)
+        elif tag_type == 'affiliation':
+            queryset = Profile.objects.filter(affiliation__id=tag_id)
+        else:
+            return Response({'message': 'Invalid tag type. Options are: skill_known, skill_available, skill_wanted, language, territory, wikimedia_project, affiliation.'}, status=status.HTTP_400_BAD_REQUEST)
+
+        return Response(self.get_serializer(queryset, many=True).data)
